Refactored PlanDate
Refactored the class PlanDate
