+++ TaskManager/src/application/DatePicker.java
@@ -1,0 +1,171 @@
+package application;
+
+import java.time.DayOfWeek;
+import java.time.LocalDate;
+import java.time.temporal.TemporalAdjusters;
+import java.time.temporal.WeekFields;
+import application.Main.ViewMode;
+import javafx.geometry.Pos;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.layout.GridPane;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.Priority;
+import javafx.scene.layout.VBox;
+
+/**
+ * Extends HBox. Holds buttons corresponding to days and weeks in a month. Buttons update the Calendarpane to show that day or week.
+ */
+public class DatePicker extends HBox {
+	CalendarPane calendarPane;
+	LocalDate pickerDate;
+	GridPane dateGrid;
+	Label dateLabel;
+
+	/**
+	 * @param calendarPane Calendarpane linked to this DatePicker will update to show tasks for the picked date.
+	 */
+	DatePicker(CalendarPane calendarPane) {
+		this.calendarPane = calendarPane;
+		//Get copy of date
+		this.pickerDate = calendarPane.getPlanDate();
+
+		//Create transparent border so taskpane is visible behind DatePicker, when clicked on transparent borders the DatePicker disappears
+		this.setStyle("-fx-background-color: transparent;");
+		this.setAlignment(Pos.TOP_CENTER);
+		this.setOnMouseClicked(e -> { this.setVisible(false); });
+
+		//Create inner pane to hold date buttons
+		VBox innerPane = new VBox();
+		this.getChildren().add(innerPane);
+		innerPane.setStyle("-fx-background-color: white; -fx-border-color: grey;");
+		innerPane.setMaxHeight(249);
+
+		//DatePicker doesn't disapear when the innerPane is clicked
+		innerPane.setOnMouseClicked(e -> { e.consume(); });
+
+		//HBox for date title buttons
+		HBox datePane = new HBox();
+
+		//Month and year label
+		this.dateLabel = new Label();
+		this.dateLabel.setMinHeight(31);
+		this.dateLabel.setMaxWidth(Double.MAX_VALUE);
+		HBox.setHgrow(dateLabel, Priority.ALWAYS);
+		this.dateLabel.setAlignment(Pos.BASELINE_CENTER);
+
+		//Previous month button
+		Button previousDayButton = new Button("<");
+		previousDayButton.setMinSize(31, 31);
+		previousDayButton.setMaxSize(31, 31);
+		previousDayButton.setAlignment(Pos.BASELINE_CENTER);
+		previousDayButton.setOnAction(e -> {
+			this.pickerDate = this.pickerDate.minusMonths(1);
+			this.Update();
+		});
+
+		//Today button
+		Button todayButton = new Button("^");
+		todayButton.setMinSize(31, 31);
+		todayButton.setMaxSize(31, 31);
+		todayButton.setAlignment(Pos.BASELINE_CENTER);
+		todayButton.setOnAction(e -> {
+			this.pickerDate = LocalDate.now();
+			this.setVisible(false);
+			this.calendarPane.setViewMode(ViewMode.day);
+			this.calendarPane.setPlanDate(this.pickerDate);
+		});
+
+		//Next month button
+		Button nextDayButton = new Button(">");
+		nextDayButton.setMinSize(31, 31);
+		nextDayButton.setMaxSize(31, 31);
+		nextDayButton.setAlignment(Pos.BASELINE_CENTER);
+		nextDayButton.setOnAction(e -> {
+			this.pickerDate = this.pickerDate.plusMonths(1);
+			this.Update();
+		});
+
+		//Add datebuttons to bar
+		datePane.getChildren().addAll(previousDayButton, todayButton, this.dateLabel, nextDayButton);
+
+		//Datepicker pane
+		this.dateGrid = new GridPane();	
+
+		//Add to inner pane
+		innerPane.getChildren().addAll(datePane, this.dateGrid);
+
+	}
+
+	/**
+	 * Updates dates on the DatePicker buttons to the month selected with the month navigation buttons on the DatePicker.
+	 */
+	void Update() {			
+		//Clear dates
+		this.dateGrid.getChildren().clear();
+
+		//Day of week headers
+		String[] dayTitleList = {"Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"};
+		for (int i = 0; i < dayTitleList.length; i++) {
+			Label dayTitleLabel = new Label(dayTitleList[i]);
+			dayTitleLabel.setAlignment(Pos.BASELINE_CENTER);
+			dayTitleLabel.setMinSize(30, 30);
+			this.dateGrid.add(dayTitleLabel, 1 + i, 0);
+		}
+
+		//Set date title
+		this.dateLabel.setText(this.pickerDate.getMonth().toString() + " " + this.pickerDate.toString().substring(0, 4));
+
+		//Initialize starting date
+		LocalDate startDate = this.pickerDate.withDayOfMonth(1).with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY));
+
+		//Create buttons for 42 dates
+		for (int i = 0; i < 42; i++) {
+			//Initialize PlanDate from a number of days after starting date
+			LocalDate followingDate = startDate.plusDays(i);
+
+			//Create button for PlanDate
+			Button dateButton = new Button(followingDate.toString().substring(8,10));
+			dateButton.setAlignment(Pos.BASELINE_CENTER);
+			dateButton.setMinSize(31, 31);
+			dateButton.setMaxSize(31, 31);
+
+			//Set button text black if it matches the currently displayed month, otherwise grey, and set border black if it matches the current date
+			if (followingDate.toString().substring(0,7).equals(this.pickerDate.toString().substring(0,7))) {
+				dateButton.setStyle("-fx-text-fill: black;");
+				if (followingDate.equals(LocalDate.now())) {
+					dateButton.setStyle("-fx-border-color: black; -fx-border-width: 1;");
+				} else {
+					dateButton.setStyle("-fx-border-color: transparent; -fx-border-width: 1;");
+				}
+			} else {
+				dateButton.setStyle("-fx-text-fill: grey;");
+			}
+
+			//Hide DatePicker and display TaskPane for date when button is clicked
+			dateButton.setOnAction(e -> {
+				this.pickerDate = followingDate;
+				this.setVisible(false);
+				this.calendarPane.setViewMode(ViewMode.day);
+				this.calendarPane.setPlanDate(this.pickerDate);
+			});
+			dateGrid.add(dateButton, 1 + i % 7, 1 + i / 7);
+
+			//Add buttons for weeks
+			if (i % 7 == 0) {
+				Button weekButton = new Button("Week " + followingDate.get(WeekFields.of(DayOfWeek.MONDAY, 4).weekOfWeekBasedYear()));
+				weekButton.setMinSize(64, 31);
+				weekButton.setMaxSize(64, 31);
+				weekButton.setAlignment(Pos.CENTER_LEFT);
+				//Hide DatePicker and display TaskPane for week when button is clicked
+				weekButton.setOnAction(e -> {
+					this.pickerDate = followingDate;
+					this.setVisible(false);
+					this.calendarPane.setViewMode(ViewMode.week);
+					this.calendarPane.setPlanDate(this.pickerDate);
+				});
+				dateGrid.add(weekButton, 0, 1 + i / 7);
+			}
+		}
+	}
+}
