--- TaskManager/src/application/PlanDate.java
@@ -1,129 +1,0 @@
-package application;
-
-import java.sql.SQLException;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.GridPane;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.Priority;
-import javafx.scene.layout.VBox;
-import javafx.scene.text.Font;
-
-/**
- * Represents a date in the agenda. Holds a list of MainTasks.
- */
-public class PlanDate {
-	
-	LocalDate date;
-	ArrayList<MainTask> taskList = new ArrayList<MainTask>();
-	VBox taskBox = new VBox();
-
-	/**
-	 * Creates PlanDate for the specified date.
-	 * @param date
-	 */
-	public PlanDate(LocalDate date) {
-		this.date = date;
-	}
-	
-	@Override
-	public String toString() {
-		return this.date.toString();
-	}
-	
-	
-	/**
-	 * Gets tasks for the date of this PlanDate from the SQLite database.
-	 * Retrieved MainTasks are stored in taskList in this PlanDate. SubTasks are stored in subTaskList in their respective MainTasks.
-	 */
-	public void getTasks() {
-		//Clear tasklist
-		this.taskList.clear();
-
-		//Get tasks from database
-		SQLConnector.read("SELECT * FROM tasks WHERE Date = '" + this.toString() + "'", rs -> {
-			try {
-				while (rs.next()) {
-					//Parse string result from time query to LocalTime if it's not null
-					LocalTime taskTime = null;
-					if (rs.getString("Time") != null) {
-						taskTime = LocalTime.parse(rs.getString("Time"));
-					}
-
-					//Create new main task
-					MainTask newMainTask = new MainTask(rs.getInt("ID"), rs.getString("Name"), this, taskTime, Boolean.parseBoolean(rs.getString("Completed")), Boolean.parseBoolean(rs.getString("Expanded")));
-
-					//Get subtasks for main task
-					SQLConnector.read("SELECT * FROM subtasks WHERE MainTaskID = " + newMainTask.getID(), rsSubTask -> {
-						try {
-							while (rsSubTask.next()) {
-								//Create subtask
-								SubTask newSubTask = new SubTask(newMainTask, rsSubTask.getInt("ID"), rsSubTask.getString("Name"), Boolean.parseBoolean(rsSubTask.getString("Completed")));
-								//Add to subtasklist
-								newMainTask.addToSubTaskList(newSubTask);
-							}
-						} catch (SQLException e) {
-							System.out.println(e);
-						}
-					});
-					
-					//Add main task to tasklist
-					this.taskList.add(newMainTask);
-				}
-			} catch (SQLException e) {
-				System.out.println(e);
-			}
-		});
-	}
-	
-	/**
-	 * Creates a UI element containing the MainTaskPane of all MainTasks in this PlanDate.
-	 * The element VBox taskBox holds the MainTaskPanes and can be updated with updateTaskBox() when tasks are changed.
-	 * @return VBox containing date label and MainTaskPanes
-	 */
-	public VBox createDayBox() {
-		VBox dayBox = new VBox();
-		dayBox.setPadding(new Insets(0, 3, 3, 3));
-		dayBox.setStyle("-fx-border-color: grey; -fx-border-width: 1;");
-		
-		dayBox.setMaxSize(Double.MAX_VALUE, Double.MAX_VALUE);
-		GridPane.setHgrow(dayBox, Priority.ALWAYS);
-		GridPane.setVgrow(dayBox, Priority.ALWAYS);
-
-		//Day of week label
-		HBox titleBox = new HBox();
-		Label dayOfWeekLabel = new Label(this.date.getDayOfWeek().toString().substring(0,3));
-		dayOfWeekLabel.setMinWidth(40);
-		Label dateLabel = new Label(this.date.format(DateTimeFormatter.ofPattern("dd-MM-yyyy")));
-		titleBox.getChildren().addAll(dayOfWeekLabel, dateLabel);
-
-		//Labels for tasks with spacing of 3 between tasks and between following add task button
-		this.taskBox = new VBox();
-		this.taskBox.setSpacing(3);
-		this.taskBox.setPadding(new Insets(0, 0, 3, 0));
-
-		dayBox.getChildren().addAll(titleBox, this.taskBox);
-
-		return dayBox;
-	}
-	
-	/**
-	 * Updates taskBox with MainTaskPanes of MainTasks in this PlanDate.
-	 */
-	public void updateTaskBox() {
-		this.taskBox.getChildren().clear();
-		this.taskList.sort(null);
-		for (MainTask task : this.taskList) {
-			this.taskBox.getChildren().add(task.taskPane);
-		}
-	}
-}
